+++ packages/r3f-perf/src/gui.ts
@@ -1,0 +1,1 @@
+export { Gui as Perf } from './perfgui';