+++ packages/r3f-perf/src/headless.ts
@@ -1,0 +1,2 @@
+export { usePerf, getPerf, setPerf, setCustomData, getCustomData } from './store';
+export { Headless as PerfHeadless } from './perfheadless';