+++ packages/r3f-perf/src/html.tsx
@@ -1,5 +1,6 @@
 import { useThree } from '@react-three/fiber'
 import React, { forwardRef, ReactNode, useLayoutEffect, useRef } from 'react'
+// @ts-ignore
 import { createRoot, Root } from 'react-dom/client'
 
 interface HtmlProps {