+++ packages/r3f-perf/src/index.ts
@@ -1,0 +1,2 @@
+export { usePerf, getPerf, setPerf, setCustomData, getCustomData } from './store';
+export { Gui as Perf } from './perfgui';